backend/requirements.txt
Flask==3.0.0
tensorflow==2.19.0
Pillow==10.0.1
flask-cors==4.0.0
numpy==1.26.0
gunicorn==21.2.0